+++ server_src/game_model/duck.cpp
@@ -65,8 +65,8 @@ bool Duck::is_this_duck(int id) { return this->duck_id == id; }
 duck_DTO Duck::to_DTO(){
     duck_DTO dto;
     dto.duck_id = this->duck_id;
-    uint8_t x = this->position.get_x();
-    uint8_t y = this->position.get_y();
+    uint16_t x = this->position.get_x();
+    uint16_t y = this->position.get_y();
     dto.x = x;
     dto.y = y;
     dto.running = false;
